pyproject.toml
[tool.poetry]
name = "mode-groothandel-synchronizer"
version = "0.1.0"
description = "A synchronizer for Mode Groothandel"
authors = ["Lars van Rhijn <lars@larsvanrhijn.nl>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
django-import-export = "^3.3.8"
django-constance = "^3.1.0"
djangorestframework = "^3.15.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pydocstyle = "^6.1.1"
flake8 = "^6.0.0"
freezegun = "^1.2.1"
tomli = "^2.0.1"

[tool.poetry.group.prod.dependencies]
uwsgi = "^2.0.23"
psycopg2-binary = "^2.9.1"
sentry-sdk = "^1.14.0"

[tool.black]
line-length = 119
target-version = ["py310"]
exclude = '''
/(
  migrations
  | website/mode_groothandel/settings
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
